backend/requirements.txt
anytree==2.8.0
appdirs==1.4.4
asgiref==3.5.2
Beaker==1.11.0
Brlapi==0.8.3
btrfsutil==5.18.1
CacheControl==0.12.6
cairocffi==1.3.0
CairoSVG==2.5.2
cffi==1.15.0
chardet==4.0.0
colorama==0.4.4
contextlib2==0.6.0.post1
cryptography==37.0.2
cssselect2==0.5.0
defusedxml==0.7.1
distlib==0.3.4
distro==1.7.0
Django==4.0.5
django-cors-headers==3.13.0
djangorestframework==3.13.1
html5lib==1.1
idna==3.3
importlib-metadata==4.8.1
jaraco.context==4.1.1
jaraco.functools==3.5.0
jaraco.text==3.8.0
lensfun==0.3.3
libfdt==1.6.1
louis==3.22.0
lxml==4.8.0
Mako==1.1.6
Markdown==3.3.7
MarkupSafe==2.1.1
meson==0.62.2
more-itertools==8.12.0
msgpack==1.0.3
netsnmp-python==1.0a1
nspektr==0.4.0
ordered-set==4.0.2
packaging==21.3
pep517==0.12.0
Pillow==9.1.1
ply==3.11
progress==1.6
psutil==5.9.0
pwquality==1.4.4
pycairo==1.21.0
pycparser==2.21
Pygments==2.12.0
PyGObject==3.42.1
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyperclip==1.8.2
python-dbusmock==0.27.5
python-decouple==3.6
python-xlib==0.31
pytz==2022.1
pyudev==0.22.0.dev20211202
pyxdg==0.28
requests==2.27.1
resolvelib==0.5.5
retrying==1.3.3
six==1.16.0
sqlparse==0.4.2
svgwrite==1.4.1
team==1.0
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
urllib3==1.26.9
webencodings==0.5.1
xcffib==0.11.1
zipp==3.8.0
